+++ index.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const parseSubprojects = require('./lib/parseSubprojects');
 const prepareContext = require('./lib/prepareContext');
 const renderHerman = require('./lib/renderHerman');
 
@@ -16,9 +15,7 @@ const preview = require('./lib/annotations/preview');
  */
 const herman = (dest, ctx) =>
   prepareContext(ctx).then(preparedContext =>
-    parseSubprojects(preparedContext).then(() =>
-      renderHerman(dest, preparedContext)
-    )
+    renderHerman(dest, preparedContext)
   );
 
 herman.annotations = [icons, preview, font, example, name];